+++ audiomate/encoding/__init__.py
@@ -6,3 +6,5 @@ from .base import Encoder  # noqa: F401
 
 from .frame_based import FrameHotEncoder  # noqa: F401
 from .frame_based import FrameOrdinalEncoder  # noqa: F401
+
+from .utterance_based import TokenOrdinalEncoder  # noqa: F401